Updated several settled rulings
* Added several settled rulings around Attack Modifier Cards to Section 5.3.

If I draw multiple Heal Ally (or Heal Self) modifier draws, do they combine into one heal, or is each one separate? They all combine into one heal ability, but you can either combine them to heal one target, or split them up to heal two or more. All heals to one target are combined into a single heal, and are not treated as multiple Heal 1’s. (This was moved from the Deathwalker section and made more general since so many classes have these.)

If I draw an attack modifier for a summon's attack that refers to something that doesn't make sense for it, for example gaining one of my tokens or a reference to one of my summons, how is this resolved? All class-specific mechanics are controlled by the summoner, even when drawn by a summon. You gain the benefits of these attack modifier cards. (This is straight from the rulebook on pg 26 but has been the subject of frequent questions from the community.)

Added 2 AMD questions to Prism section.