clarify negative item effects wording
The original reads as though the choice not to equip is dependent on not
having the relevant perk, which is not the intended meaning. It also
reads as though gaining the -1s are required if you equip the item,
which is also not quite the intended meaning (since those -1s are
avoided by having the perk!).

Instead, note that you can _always_ choose not to equip an item. Adding
the -1s is contingent on equipping the item sans relevant perk.